Modifcacion de entidades

+++ Desguambios/src/main/java/com/dsg/desguambios/entidades/Producto.java
@@ -11,7 +11,7 @@ public class Producto {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private long id_producto;
+	private Integer id_producto;
 	
 	
 	private String lit_producto;
